refactor: remove ConsoleExtensions
